fix(ci): CI全ジョブ緑化・ペガサス500行ルール対応 (Issue #126)
【CI修正】
- cargo fmt --all: Mayu追加テストのフォーマット違反を修正
  (cli.rs:521,528 / lib.rs:403,432 / e2e_sc_ddm_001.rs / it_cli_default_mode.rs)
- tc_f_i11b: --no-ipc追加（IPC既定化後のパストラバーサル検証経路修正）
  DaemonNotRunning(exit 1)でなくPersistence(exit 2)を正しく検証するため
- smoke-e2e.sh: `shikomi list --ipc` → `shikomi list`（--ipcフラグ廃止）
- test-gui.yml: fault-check も同様に --ipc を除去
- audit-secret-paths.sh: TC-CI-026 allowlistにtests.rsを追加
  (TC-UT-155の unsafe {set_var} がlib.rs本番コードに混入しないことを保証)

【ペガサス 500行ルール対応】
- cli.rs (539→291行): #[cfg(test)]ブロックをsrc/cli_tests.rsに分離
  #[path = "cli_tests.rs"] mod tests; で参照
- lib.rs (510→400行): #[cfg(test)]ブロックをsrc/tests.rsに分離
  mod tests; で参照（src/ディレクトリ直下のため#[path]不要）
