+++ .github/workflows/test-gui.yml
@@ -118,5 +118,6 @@ jobs:
 
       - name: fault check (IPC must fail when daemon is not running)
         # 設計根拠: e2e.md §6.8 ステップ 6
-        # daemon 未起動 → shikomi list が exit 非ゼロ → ! で反転 → ステップ PASS
-        run: "! ./target/release/shikomi list --ipc"
+        # daemon 未起動 → shikomi list が exit 非ゼロ（DaemonNotRunning）→ ! で反転 → ステップ PASS
+        # NOTE: Issue #126 で IPC 既定化済み。`--ipc` フラグは廃止。
+        run: "! ./target/release/shikomi list"
